Documentation and fixed some issues
I was adding some basic documentation for the whomever may need it, and
realized I had forgotten to test for "get_all". Doing that led me to a
whole host of new issues that I had to fix. The good news is that the
REST client now returns all data correctly. The bad news is that the
server throws an exception when trying to insert a duplicate username,
and I have not yet found a way to catch it. This translates to a status
code 500 in the REST client which it doesn't know how to handle and
throws an exception. So you'll see in the unit test for testing
duplicate inserts that it works-ish by way of an exception, but doesn't
return any type of valid JSON or ID indicator. Otherwise, vast
improvements from even yesterday.